+++ src/main/java/christmas/exception/order/OrderDuplicateMenuException.java
@@ -1,4 +1,4 @@
-package christmas.exception;
+package christmas.exception.order;
 
 public class OrderDuplicateMenuException extends IllegalArgumentException {
     private static final String MESSAGE =  "[ERROR] 유효하지 않은 주문입니다. 다시 입력해 주세요.";
