+++ src/test/java/christmas/domain/order/OrdersTest.java
@@ -6,10 +6,10 @@ import static org.junit.jupiter.api.Assertions.assertThrows;
 import christmas.domain.VisitDate;
 import christmas.exception.DateUnitException;
 import christmas.exception.NotExistsMenuException;
-import christmas.exception.OrderDuplicateMenuException;
-import christmas.exception.OrderOnlyDrinkException;
-import christmas.exception.OrderTotalExceededException;
-import christmas.exception.OrderValueMinUnitException;
+import christmas.exception.order.OrderDuplicateMenuException;
+import christmas.exception.order.OrderOnlyDrinkException;
+import christmas.exception.order.OrderTotalExceededException;
+import christmas.exception.order.OrderValueMinUnitException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
